+++ internal/repository/postgres/users.go
@@ -5,7 +5,7 @@ import (
 	"final/internal/repository/postgres/sqlc"
 )
 
-func (p *Postgres) CreateUser(ctx context.Context, arg sqlc.CreateUserParams) (int32, error) {
+func (p *Postgres) CreateUser(ctx context.Context, arg sqlc.CreateUserParams) (int64, error) {
 	// Логируем, чтобы убедиться, что статус реально передаётся
 	p.logger.Debug("CreateUser called", "login", arg.Login, "status", arg.Status)
 
@@ -18,7 +18,7 @@ func (p *Postgres) GetUserByLogin(ctx context.Context, login string) (*sqlc.User
 }
 
 func (p *Postgres) GetUserByID(ctx context.Context, id int64) (*sqlc.User, error) {
-	return p.queries.GetUserByID(ctx, int32(id))
+	return p.queries.GetUserByID(ctx, id)
 }
 
 func (p *Postgres) UpdateUserStatus(ctx context.Context, arg sqlc.UpdateUserStatusParams) error {
@@ -30,9 +30,9 @@ func (p *Postgres) UpdateUserName(ctx context.Context, arg sqlc.UpdateUserNamePa
 }
 
 func (p *Postgres) DeleteUser(ctx context.Context, id int64) error {
-	return p.queries.DeleteUser(ctx, int32(id))
+	return p.queries.DeleteUser(ctx, id)
 }
 
-func (p *Postgres) Logout(ctx context.Context, id int32) error {
+func (p *Postgres) Logout(ctx context.Context, id int64) error {
 	return p.queries.Logout(ctx, id)
 }
